requirements.txt
docker
prompt-toolkit
pyyaml
requests